+++ package.json
@@ -24,6 +24,7 @@
   },
   "author": "Olga Stefanishyna",
   "dependencies": {
+    "bcrypt": "^3.0.7",
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
